Finished PL.Test Code for utDirector, ut Format, utMovie, utRating, utTableStructure. All Passed

+++ DDB.DVDCentral.DB/Tables/tblMovie.sql
@@ -3,7 +3,7 @@
 	[Id] INT NOT NULL PRIMARY KEY,
 	[Title] varchar(50) NOT NULL,
 	[Description] varchar(250) NOT NULL, 
-    [Cost] MONEY NOT NULL,
+    [Cost] FLOAT NOT NULL,
 	[RatingId] INT NOT NULL,
 	[FormatId] INT NOT NULL,
 	[DirectorId] INT NOT NULL,
